When using /var/yp/securenet the local host running ypserv wasn't garanteed access to ypserv. This has been fixed by always allowing the host 127.0.0.1 access when using /var/yp/securenet. This to keep compability with SunOS.
When using the ypserv.acl(5) format access must be granted to 127.0.0.1 in
the file. This to make it possible to disallow access to ypserv from the
host running it. This has been documented in man page and example.

Thanks to <Ulrich.Flegel@braunschweig.netsurf.de> for bringing this to
my attention (again). -moj