Major changes here and there, tweaks elsewhere. Support for Openfirmware drivers was reintroduced so that more systems were supported. This should work with the real driver configurations as well.
Bootloader files were deleted/replaced with the newer versions in the
subdirectory. Some effort has been made to be closer to support booting
(at least the bootloader) on the Mac.

Config files that end with OFW are the openfirmware versions of the kernels
without have native drivers.

Native driver support has not been changed, presumably it still works.
I couldn't test that.