Escape stuff in the peers accmap as well as our own when async encoding, and back out the negotiation strategy that didn't quite have the right effect with some (bad?) ppp implementations.
Drop unexpected PAP/CHAP packets (if not in NETWORK or ESTABLISH
phase).

Make sure that we enter PHASE_ESTABLISH when we need to.

Detect HDLC frames correctly, not just when the whole header
is obtained by a single read().

Don't convert \r\n to \r\r\n in tty raw mode.

Use the correct macros for `', ``'', () and [].