Recursive definitions are not the real problem. Trying to expand them is.
So flag obvious recursive definitions for later, and give an error
only if we expand them.

(Some gnu-m4 files, including autoconf, do define some macros with
themselves as the replacement text, for use in test-if-set patterns)

Since type is no longer MACRTYPE, those macros end up in builtins...
but this is not a problem, since expanding them is an error.