Add driver for Prolific PL2301/PL2302 host-to-host adapter. From NetBSD. /* XXX - untested */
Note: just reminding myself that I need to create a nice diff for Lennart
to further portablize the USB code (I fell behind in doing this). In
particular, for this driver, the IF_INPUT() macro just does ether_input()
in OpenBSD, but we really want to call the local upl_input() routine.