Improve the page mapped check algorithm in the powerpc pmap module, before it was looking through two arrays of 8 and a linked list of undetermined size, before deciding that a mapping was not valid. Now it allocates a data structure and caches that data.
This improves both pmap_enter and pmap_remove because both check
to see if a mapping is valid before taking the appropriate actions.

Also in pmap_remove, if the va mapping is found, stop searching for
it in the rest of this array, the alternate array and the linked list.
only one valid mapping of each va is allowed.

This change improved lat_mmap (from lmbench) from 1300 to 720
and fork+exit from 7320 to 2724 microseconds.