The "powerpc" port which has supported the newer Apple Macintosh powerpc based is being renamed to macppc. This is to allow sharing of common code between different powerpc base platforms.
Most of the work involved in the renaming process was performed by miod@

Files moved from powerpc/powerpc to macppc/macppc

This moves hardware specific files from the common directory to the
platform specific directory. This leaves common files.
With this change all of the debugger (db_) files have been moved to
the platform specific directory. The debugger should be reconsidered
and commonized.