Remove setuid root from lp*.  lpr needs to be setuid daemon so the files it creates are not owned by the user spooling them but the others (lpc, lpq, lprm) can get away with setgid daemon.  lpd runs as user daemon for most things, only changing its uid to 0 for things that must be done as root.
For the time being, don't require connections to come from a reserved
port since lpq/lpr/lprm can't acquire that w/o setuid root.  In the
near future we will have a mechanism for select non-root processes
to grab reserved ports.

The upshot of this is that spool directories must be writable by
group daemon and the files within the spool dirs must be owned by
daemon.