Add 'option NFSCLIENT' to macppc RAMDISK.
Delete (comment out) 'option NFSCLIENT' in sparc, sparc64, and sun3
RAMDISKs.

This brings the presence of /sbin/mount_nfs in the boot image and the
presence of NFSCLIENT in the RAMDISK kernel into alignment. Except for
sparc64 which will now have mount_nfs removed from the boot image to
save space.

mvme88k RAMDISK did not need to be adjusted as it only *looks* like it
does not have NFSCLIENT. It actually includes the system GENERIC.

Investigation for inconsistancies prompted by todd@.

This solution recommended by deraadt@.