Add a sysctl to option USER_LDT, to control its behaviour, which will be disabled by default.
Enable with sysctl -w machdep.userldt=1

This will allow people to use the few ports that require this functionality
to work without requiring the user to recompile a kernel.

The option USER_LDT remains in order to not increase size on the
installation media. It is now enabled in GENERIC.

Per espie@'s idea, but my diff was much simpler than his; ok deraadt@