Add a simple driver for the SBus ``Prestoserve'' NVRAM cards.
The Legato Prestoserve NFS accelerator used to be very popular, back when
NFSv3 did not exist, and disks and network speeds were not as good as they
are now...

This driver does not attempt to play any game with filesystems, but will only
provide access to the cards NVRAM, as a block device.

If you have bright ideas on ``what to do with a few MB of battery-backed
memory'', I am interested in realistic suggestions!