More cleanup and fixes in url installs.
1) Filter appropriate lines (ftp or http) from ftplist once when it is
downloaded rather than everytime the lines are needed.

2) Don't display the 'ftp://' or 'http://' part of the listed url's as
all displayed url's are of the same type. Partially compensates for
the last increase in the width of the 'cat -n' line number field.

3) Store filtered lines in /tmp/serverlist instead of /tmp/ftplist,
since it can be a list of http servers too. Makes it more obvious that
it is not a copy of ftplist.

4) Avoid 'ftplist: No such file' message, but add more informative
'Getting...done.' or 'Getting...FAILED.' message. Display the actual
ftp error that caused the FAILED status.

5) Add 'done' to the list of available options when asking for a
server location, allowing the user to return to set location selection
without having to enter a server.

6) Mention the '?' option in the prompt. It re-displays the list of
servers.

7) Eliminate unneeded variables 'host' and 'maxlines'. Make global
tline a local (_line). Change _ftp_get_list global to more generic
_get_server_list.

8) General logic cleanup.