Our extent implementation relies on the fact that extents all have distinct names.
However, the tsp code uses one global extent, which both device instances
want to reinitialize and use for themselves (whether this is a genuine bug
or not is left to the reader).

Solve the problem by putting the extents into the tsp softc, and change the
pci_bwx_bus code to handle variable extent names.

This makes DS20 systems with tsp0 and tsp1 work again.

ok deraadt@