MD installation notes updates for 3.4, 3/3
Describe with much more details how to successfully boot from various
devices, including non-built-in ethernet or SCSI controllers; also hint
that most of the MVME147 and the MVME187 can not netboot at all.

This should be much, much more understandable (but we need to run this
through a drunk pvalchev@ to be sure).