gcc 3.3.5 loop optimization will produce incorrect output on powerpc; a victim of this being awk(1), which in turn caused some manual pages to be generated without dates at the bottom.
Since this is fixed by the (invasive) loop optimization rewrite done for
3.4, we can not expect a 3.3 fix anytime soon, and the fix is non-trivial,
so disable the failing section of code on powerpc only.

This has the net effect of magically turning -O1 and -O2 into -Os as far as
-floop-optimize is concerned.

Problem spotted by deraadt@, researched by otto@ and yours truly; this
workaround agreed by many.