Start on a basic ACPI framework -- does not do much more than read out the ACPI tables into kernel memory and attach ACPI and HPET timers currently.
In order to test this code, enabling the devices in GENERIC as well as
the ACPI_ENABLE option is needed.  This code does not do any thermal
control yet, so this should be done with care depending on the platform.

In the tree so more people can contribute to making this more fully
featured.

Ok niklas@ grange@ tedu@