import sili(4) so it can be worked on in the tree. this will support the silicon image 3124/3132/3531 sata chipsets eventually.
so far this is the autoconf glue, an actual match routine for the 3124, and
interrupt establishment code. it is split up between pci and ic cos there
are cardbus variants of these controllers that we can support in the
future.

thanks to jolan@ for the name, it was better liked than my initial
suggestion of siisl(4).