While we are determining which NCQ command failed, we will need to ensure that we don't accidentally complete any other outstanding commands.
This introduces wrappers around get_ccb and put_ccb that can temporarily
stash our outstanding command state while we are issuing recovery commands.

This is just the first step in NCQ recovery - for now we will reset the port
on error which will permit further commands to be issued after a NCQ error.

ok dlg@