jmicron controllers are strange cos they provide both ide and sata support on the same pci function. they can be configured to put the different support on different pci functions after booting, but that is unreliable since the kernel doesnt know enough to be able to configure the interrupts properly on this new pci function.
this adds a "nexus" driver called jmb that will allow two drivers to attach
to the one pci device. this code already fixes things so ahci can attach to
the sata stuff. this is more reliable than switch functions on since the
interrupts on function 0 are configured by the firmware/bios/whatever.