when looping through an array and advancing your pointer to the array 2x on each loop, only loop .5x times
fixes hard hangs on i386 and "hiccups" on sgi when these functions are
used through e.g. dd if=/dev/audio of=foo, since mulaw is the default
encoding.

the commit history wrt this code is interesting.  I am responsible for
introducing this braindamage to OpenBSD ... and note that
linear16_to_linear8_[lb]e had it right ...