Isakmpd acquire mode did not work with a config generated from ipsec.conf.  The config created by isakmpd dynamically was different from the config that ipsecctl generated out of ipsec.conf.
Both config formats are changed so that they match.  One needs a
passive ike line and a require flow line with the same parameters
in the ipsec.conf.  Then the acquire message generated by the kernel
will trigger isakmpd to generate a config that matches the one that
ipsecctl generated from the ike line.

ok hshoexer, 'sounds good' todd