Add support for the ``Firefox'' VAXstation 3520/3540/3820/3840 workstations, currently limited to serial console and a single processor working.
All ``on-board'' devices, including the Q-bus adapter, but except for
the frame buffer, are supported. The machine will boot over the network
or from SCSI devices.

Lots of thanks to Al Kossow for www.bitsavers.org, on which I found the
technical documentation allowing me to complete this port (which was
lacking at the time I got that machine...).