when separating command from parameters in smtp session, the parser tries to use ':' as a separator then fallbacks to ' ' so that it can detect the command names that contain more than one words (MAIL FROM and RCPT TO) or the one word ones (HELO, DATA, ...).
this is incorrect and the parser can get confused if the parameter to any
command contains a ':', for example "HELO [ipv6:...]" cause the parser to
lookup for command "HELO [ipv6".

fix this by using ':' as a delimiter for 'mail from' and 'rcpt to', while
using ' ' as a delimiter for all other commands.

fixes bug 6285/system reported by Lionel Le Folgoc <lionel@lefolgoc.net>