Align FFS filesystem start and end sectors so the start is on a bsize boundary and the end fills up the last bsize chunk.
Don't change the start sector if it is the first sector of the
OpenBSD portion of the disk.

Don't attempt to align on SUN_CYLCHECK architectures. They are
attempting a different alignment.

This is an attempt to ensure that FFS i/o's are aligned for optimal
performance on newer disks that lie about their sector size.