new check in pkg_create: walk the dependency tree for @wantlib, thus preventing packages that won't install (much easier to fix that when you still have the fake tree around !)
specifically, look into depend that are installed, and possibly in the
fake tree: if the package does build, it requires the libraries to be
around anyways, so pure RUN_DEPENDS are not an issue (and avoid old libraries,
if we find wantlib only in .lib, something really fishy is going on).

Reuse the Dependencies class, so we're sure to have the exact same way
to find libraries.

Also fix exit codes from pkg_create...